# docs/swagger.json
{ "openapi": "3.0.0",
    "info":{
        "title":"Speckio Web App",
        "description":"This is an express API for the Speckio web app. Made by NHMG Development.",
        "version":"1.0.0"
    },
    "servers":[
        {
            "url": "https://localhost"
        }
    ],
    "components": {
      "securitySchemes": {
        "ApiKeyAuth": {
          "type": "apiKey",
          "in": "header",
          "name": "authorization"
        }
      }
    },
    "security": [
      { "ApiKeyAuth": [] }
    ],
    "paths":{
...
      "/org/admin": {
        "get": {
          "tags": [
            "Organisations"
          ],
          "description": "Return an array of admins in organisations",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "team": {
                        "type": "array",
                        "example": [
                          {
                            "userID": 1,
                            "firstName": "John",
                            "lastName": "Smith",
                            "email": "johns@speckio.com"
                          },
                          {
                            "userID": 3,
                            "firstName": "George",
                            "lastName": "Adams",
                            "email": "georgea@speckio.com"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "Admins not found"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Organisations"
          ],
          "description": "Add user to admins",
          "parameters": [
            {
              "name": "userID",
              "in": "query",
              "description": "ID of user",
              "required": true,
              "schema": {
                "type": "integer",
                "example": "3"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "User added to admins"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "Unable to add user to admins"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Organisations"
          ],
          "description": "Delete a user from admins",
          "parameters": [
            {
              "name": "userID",
              "in": "query",
              "description": "ID of user",
              "required": true,
              "schema": {
                "type": "number",
                "example": 3
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "team": {
                        "type": "object",
                        "example": {
                          "message": "User removed from admins"                        
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "Unable to remove admin"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
}